backend/requirements.txt
flask>=2.3.0
flask-sqlalchemy>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.8.0
python-engineio>=4.7.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
facenet-pytorch>=2.5.0
pillow>=10.0.0
requests>=2.31.0
email-validator>=2.0.0
pytest>=7.4.0
pytest-flask>=1.2.0